feat(linter): Split jest `valid-describe-callback` rule
