+++ crates/oxc_linter/data/vitest_compatible_jest_rules.json
@@ -19,6 +19,5 @@
   "prefer-to-have-been-called-times",
   "prefer-to-have-length",
   "require-hook",
-  "require-to-throw-message",
-  "valid-describe-callback"
+  "require-to-throw-message"
 ]
