+++ crates/oxc_linter/src/rules/shared/jest_vitest/mod.rs
@@ -21,6 +21,7 @@ pub mod prefer_expect_assertions;
 pub mod prefer_to_contain;
 pub mod prefer_todo;
 pub mod require_top_level_describe;
+pub mod valid_describe_callback;
 pub mod valid_expect;
 pub mod valid_expect_in_promise;
 pub mod valid_title;
